requirements.txt
open3d
trimesh
pyrender
Pillow
image-reward
openai
backoff
salesforce-lavis
tqdm
matplotlib